Motifs of network hierarchy

net_x_hierarchy() collects the measures of hierarchy into a single motif, which can be used to compare the relative hierarchy of different networks. The measures of hierarchy are:

net_by_connectedness() measures the proportion of dyads in the network that are reachable to one another, or the degree to which network is a single component.

net_by_efficiency() measures the Krackhardt efficiency score.

net_by_upperbound() measures the Krackhardt (least) upper bound score.

net_by_reciprocity() measures the proportion of ties in the network that are reciprocated, which is a measure of the degree to which the network is non-hierarchical.

net_x_hierarchy(.data)

Arguments

.data

A network object of class mnet, igraph, tbl_graph, network, or similar. For more information on the standard coercion possible, see manynet::as_tidygraph().

Value

A network_motif named numeric vector or sometimes a data frame with one row and a column for each motif type, giving the count of each motif in the network. This is printed as a tibble to avoid greedy printing of long vectors.

References

On hierarchy

Krackhardt, David. 1994. Graph theoretical dimensions of informal organizations. In Carley and Prietula (eds) Computational Organizational Theory, Hillsdale, NJ: Lawrence Erlbaum Associates. Pp. 89-111.

Everett, Martin, and David Krackhardt. 2012. “A second look at Krackhardt's graph theoretical dimensions of informal organizations.” Social Networks, 34: 159-163. doi:10.1016/j.socnet.2011.10.006

See also

Examples

net_x_hierarchy(ison_networkers)
#> # A tibble: 1 × 4
#>   Connectedness InvReciprocity Efficiency LeastUpperBound
#>           <dbl>          <dbl>      <dbl>           <dbl>
#> 1             1          0.209     0.0705               1