Marking nodes as core or periphery

node_is_core() identifies whether nodes belong to the core of the network, as opposed to the periphery.

node_is_core(.data, centrality = c("degree", "eigenvector"))

Arguments

.data

A network object of class mnet, igraph, tbl_graph, network, or similar. For more information on the standard coercion possible, see manynet::as_tidygraph().

centrality

Which centrality measure to use to identify cores and periphery. By default this is "degree", which relies on the heuristic that high degree nodes are more likely to be in the core. An alternative is "eigenvector", which instead begins with high eigenvector nodes. Other methods, such as a genetic algorithm, CONCOR, and Rombach-Porter, can be added if there is interest.

Value

A node_mark logical vector the length of the nodes in the network, giving either TRUE or FALSE for each node depending on whether the condition is matched.

Core-periphery

This function is used to identify which nodes should belong to the core, and which to the periphery. It seeks to minimize the following quantity: $$Z(S_1) = \sum_{(i<j)\in S_1} \textbf{I}_{\{A_{ij}=0\}} + \sum_{(i<j)\notin S_1} \textbf{I}_{\{A_{ij}=1\}}$$ where nodes \(\{i,j,...,n\}\) are ordered in descending degree, \(A\) is the adjacency matrix, and the indicator function is 1 if the predicate is true or 0 otherwise. Note that minimising this quantity maximises density in the core block and minimises density in the periphery block; it ignores ties between these blocks.

References

On core-periphery partitioning

Borgatti, Stephen P., & Everett, Martin G. 1999. "Models of core /periphery structures". Social Networks, 21, 375–395. doi:10.1016/S0378-8733(99)00019-2

Lip, Sean Z. W. 2011. “A Fast Algorithm for the Discrete Core/Periphery Bipartitioning Problem.” doi:10.48550/arXiv.1102.5511

See also

Examples

node_is_core(ison_adolescents)
#>   Betty Sue   Alice Jane  Dale  Pam   Carol Tina 
#> 1 FALSE TRUE  TRUE  FALSE TRUE  FALSE FALSE FALSE
ison_adolescents %>% 
   mutate(corep = node_is_core())
#> 
#> ── # The Adolescent Society ────────────────────────────────────────────────────
#> # A labelled, undirected network of 8 adolescents and 10 friendships ties
#> 
#> ── Nodes 
#> # A tibble: 8 × 2
#>   name  corep     
#>   <chr> <node_mrk>
#> 1 Betty FALSE     
#> 2 Sue   TRUE      
#> 3 Alice TRUE      
#> 4 Jane  FALSE     
#> 5 Dale  TRUE      
#> 6 Pam   FALSE     
#> # ℹ 2 more rows
#> 
#> ── Ties 
#> # A tibble: 10 × 2
#>    from    to
#>   <int> <int>
#> 1     1     2
#> 2     2     3
#> 3     3     4
#> 4     2     5
#> 5     3     5
#> 6     4     5
#> # ℹ 4 more rows
#>