Memberships in core-periphery categories

node_in_core() categorizes nodes into two or more core/periphery categories based on their coreness.

node_in_core(.data, groups = 3, cluster_by = c("bins", "quantiles", "kmeans"))

Arguments

.data

A network object of class mnet, igraph, tbl_graph, network, or similar. For more information on the standard coercion possible, see manynet::as_tidygraph().

groups

Number of categories to create. Must be at least 2 and at most the number of nodes in the network. Default is 3.

cluster_by

Method to use to create the categories. One of "bins" (equal-width bins), "quantiles" (quantile-based bins), or "kmeans" (k-means clustering). Default is "bins".

Value

A node_member character vector the length of the nodes in the network, of group memberships "A", "B", etc for each node. If the network is labelled, then the assignments will be labelled with the nodes' names.

Core-periphery categories

This function categorizes nodes based on their coreness into a specified number of groups. The groups are labeled as "Core", "Semi-core", "Semi-periphery", and "Periphery" depending on the number of groups specified. The categorization can be done using different methods: equal-width bins, quantile-based bins, or k-means clustering.

References

On core-periphery categorization

Wallerstein, Immanuel. 1974. "Dependence in an Interdependent World: The Limited Possibilities of Transformation Within the Capitalist World Economy." African Studies Review, 17(1), 1-26. doi:10.2307/523574

See also

Examples

node_in_core(ison_adolescents)
#> 3 groups
#>   Betty     Sue   Alice Jane           Dale  Pam            Carol     Tina     
#> 1 Periphery Core  Core  Semi-periphery Core  Semi-periphery Periphery Periphery